requirements.txt
PyYAML~=6.0
psutil~=5.9.4